refactor:#205 클라이언트 API 및 게시글 카드 컴포넌트 리팩토링 - GonggamPagination 컴포넌트를 기본 내보내기로 변경 - GonggamPostCard에서 EachPaginatedGonggamPost를 GonggamPostDetailResponse로 수정 - GonggamPostContent에서 작성자 정보를 업데이트하고 API 경로 수정 - GonggamDetail API 파일 삭제 및 관련 코드 정리 - GonggamPostDetail 인터페이스 수정 및 게시글 세부 정보 구조 변경
