+++ src/app/gonggam/(board)/[category]/page.tsx
@@ -2,11 +2,11 @@
 
 import Link from 'next/link';
 import { useSearchParams } from 'next/navigation';
-import { GonggamPagination } from '@/components/features/gonggam/gonggam-pagination';
+import GonggamPagination from '@/components/features/gonggam/gonggam-pagination';
 import GonggamPostCard from '@/components/features/gonggam/gonggam-post-card';
+import { usePaginatedGonggamPosts } from '@/lib/hooks/queries/use-paginated-gonggam-posts';
 import { slugToCategory } from '@/constants/gonggam-category';
 import { PATH } from '@/constants/page-path';
-import { usePaginatedGonggamPosts } from '@/lib/hooks/queries/use-paginated-gonggam-posts';
 
 interface GonggamCategoryBoardProps {
   params: { category: string };
