+++ src/lib/utils/api/gonggam/gonggam-detail-client.api.ts
@@ -1,5 +1,7 @@
 import { supabase } from '@/lib/utils/supabase/supabase-client';
-import type { CommentWithUser, UploadGonggamCommentParams } from '@/types/gonggam';
+import { TABLE } from '@/constants/supabase-tables-name';
+import type { TableGonggamPosts } from '@/types/supabase-const';
+import type { CommentWithUser, UploadGonggamCommentParams, GonggamPostDetail } from '@/types/gonggam';
 
 /** getLikeCountClient
  * 클라이언트에서 특정 게시글의 좋아요 수를 조회하는 함수입니다.
@@ -89,4 +91,48 @@ export const likePost = async ({ postId }: { postId: number }) => {
 export const dislikePost = async ({ postId, userId }: { postId: number; userId: string }) => {
   const { error } = await supabase.from('likes').delete().eq('post_id', postId).eq('user_id', userId);
   if (error) throw new Error(error.message);
+};
+
+/** getGonggamPostDetailByClient
+ * postId를 기반으로 공감 게시글 내용을 조회합니다.
+ *
+ * @param postId - 게시글 고유 ID
+ * @returns 게시글 본문, 이미지 배열, 작성자 정보 등 포함된 상세 정보
+ */
+export const getGonggamPostDetailByClient = async (postId: TableGonggamPosts['id']): Promise<GonggamPostDetail> => {
+  // 1. 게시글 + 작성자 조인 조회
+
+  const { data: post, error: postError } = await supabase
+    .from(TABLE.GONGGAM_POSTS)
+    .select(
+      `
+      *,
+      users (
+        nickname,
+        profile_image
+      )
+    `
+    )
+    .eq('id', postId)
+    .single();
+
+  if (postError || !post) throw new Error(postError.message);
+
+  // 2. 이미지 배열 직접 조회
+  const { data: imageData, error: imageError } = await supabase
+    .from(TABLE.GONGGAM_POST_IMAGE_PATH)
+    .select('image_url')
+    .eq('post_id', postId);
+
+  if (imageError || !imageData) throw new Error(imageError.message);
+  const images = !imageError && imageData ? imageData.map((item) => item.image_url) : [];
+
+  return {
+    ...post,
+    writer: {
+      nickname: post.users.nickname,
+      profile_image: post.users.profile_image
+    },
+    images
+  };
 };
